Add route filter and remap helper
Closes https://github.com/seamapi/typed-axios/issues/9.
